Add $HOME/.bin at the end of PATH
Prepending $HOME/.bin to $PATH can lead to some interesting issues if a
program with the same name as a system program is moved here, because
unless the user is aware of this, it can happen that a different binary
is run.

Malicious agents could even exploit this by downloading a shady
executable into $HOME/.bin with the same name as a system binary such as
ssh or vim and cause these commands to run unauthorized code.